requirements.txt
torch==1.7.0
transformers==4.0.0
pyyaml
tqdm
